pyproject.toml
[tool.poetry]
name = "anatomic"
version = "0.1"
description = "Anotomy for u"
authors = ["dx0naclyoo <konyzhenkovdanil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
redis = "^5.0.4"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
